+++ src/main/java/com/example/AgentDead.java
@@ -1,0 +1,7 @@
+package com.example;
+
+final class AgentDead implements AgentEvent {
+    final Agent killer;
+    final Agent dead;
+    AgentDead(Agent killer, Agent dead) { this.killer = killer; this.dead = dead; }
+}
