+++ src/main/java/com/example/AgentEvent.java
@@ -1,0 +1,3 @@
+package com.example;
+
+sealed interface AgentEvent permits AgentDead { }
