+++ src/main/java/com/example/Boost.java
@@ -1,0 +1,11 @@
+package com.example;
+
+final class Boost {
+    final double factor;
+    final long end;
+    Boost(double factor, long durationMs) {
+        this.factor = factor;
+        this.end = System.currentTimeMillis() + durationMs;
+    }
+    boolean expired() { return System.currentTimeMillis() > end; }
+}
